% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brca_expr}
\alias{brca_expr}
\title{some data to test}
\format{
An object of class \code{data.frame} with 305 rows and 44 columns.
}
\usage{
brca_expr
}
\description{
some data to test
}
\keyword{a}
\keyword{data}
\keyword{set}
