% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_matrix.R
\name{cor_matrix}
\alias{cor_matrix}
\title{This function provides methods for finding correlations and constructing them into conforming matrices.}
\usage{
cor_matrix(RBP_matrix, Events_matrix, method = "pearson")
}
\arguments{
\item{RBP_matrix}{RBP expression matrix}

\item{Events_matrix}{Events splicing matrix}

\item{method}{"pearson" or "spearman"}
}
\value{
The result include correlation coefficient and significance level matrix.
}
\description{
This function provides methods for finding correlations and constructing them into conforming matrices.
}
\examples{
a<-matrix(rnorm(25),nrow=5,ncol=5)
rownames(a)<-paste("gene",1:5,sep="")
b<-matrix(rnorm(25),nrow=5,ncol=5)
rownames(b)<-paste("event",1:5,sep="")
c<-cor_matrix(a,b)
head(c[1:5,])
}
