% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_matrix.R
\name{cor_method}
\alias{cor_method}
\title{This function is used to provide a method for finding correlation.}
\usage{
cor_method(x, y, method = "pearson")
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}

\item{method}{"pearson" or "spearman"}
}
\value{
The result include correlation coefficient and significance level.
}
\description{
This function is used to provide a method for finding correlation.
}
\examples{

a<-rnorm(10)
b<-rnorm(10)
cor_method(a,b,"pearson")
}
