% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_spearman.R
\name{cor_spearman}
\alias{cor_spearman}
\title{This This function provides methods for finding correlations and constructing them into conforming matrices.}
\usage{
cor_spearman(x, y, method = "spearman")
}
\arguments{
\item{x, y}{numeric data.frame or matrix. x and y must have the same dimension.}

\item{method}{a character string indicating which correlation coefficient is to be used for the test. One of "pearson", "kendall", or "spearman", can be abbreviated.}
}
\value{
The returned correlation matrix is spoofed with a total of 4 columns, the first column is the row name of the input x, the second is the row name of the input y, the third column is the magnitude of the correlation, and the fourth column is the significance level p-value. If x and y do not have row names, the returned value is the row to which they belong.
}
\description{
This This function provides methods for finding correlations and constructing them into conforming matrices.
}
\examples{
x <- matrix(sample(1:100,100),10,10)
y <- matrix(sample(1:100,100),10,10)
rownames(x)=rownames(y)=paste0("row",1:10)
colnames(x)=colnames(y)=paste0("col",1:10)
cor_spearman(x,y)
}
