% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_spearman_percent.R
\name{cor_spearman_percent}
\alias{cor_spearman_percent}
\title{This function provides the method to find the correlation based on the samples with high and low expressions in each row, taking a certain proportion and constructing it into a matrix that meets the requirements.}
\usage{
cor_spearman_percent(x, y, method = "spearman", num1 = num1, threads = 2)
}
\arguments{
\item{x, y}{numeric data.frame or matrix. x and y must have the same dimension.}

\item{method}{a character string indicating which correlation coefficient is to be used for the test. One of "pearson", "kendall", or "spearman", can be abbreviated.}

\item{num1}{Denotes the proportion/number of samples to take the expression high or low and to construct the regulatory relationship, respectively.
Input requirements: if the input is less than 1, the input is considered as the sample proportion, and greater than 1, the input is considered as the number of samples.
When choosing the input sample proportion, it is required to be no greater than 0.5.}

\item{threads}{If you want to use multiple threads, you can set threads, where the default parameter is 2.}
}
\value{
The returned correlation matrix is spoofed with a total of 4 columns,
        the first column is the row name of the input x,
        the second is the row name of the input y,
        the third column is the magnitude of the correlation,
        the fourth column is the significance level p-value.
}
\description{
This function provides the method to find the correlation based on the samples with high and low expressions in each row, taking a certain proportion and constructing it into a matrix that meets the requirements.
}
\examples{
x <- matrix(sample(1:10000,10000),100,100)
y <- matrix(sample(1:10000,10000),100,100)
rownames(x)=rownames(y)=paste0("row",1:100)
colnames(x)=colnames(y)=paste0("col",1:100)
cor_spearman(x,y)
}
