% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_integration_preprocess_rmats.R
\name{data_integration_preprocess_rmats}
\alias{data_integration_preprocess_rmats}
\title{data_integration_preprocess_rmats}
\usage{
data_integration_preprocess_rmats(
  path_rmats,
  type = "default",
  design = NULL,
  sample_num,
  psi_cutoff,
  reads_cutoff = NULL,
  reads_num = 10,
  colname_ann = FALSE,
  gsub_pattern_up = NULL,
  gsub_pattern_down = NULL
)
}
\arguments{
\item{path_rmats}{path}

\item{type}{type}

\item{design}{design}

\item{sample_num}{number}

\item{psi_cutoff}{psi_cutoff}

\item{reads_cutoff}{reads_cutoff}

\item{reads_num}{reads_num}

\item{colname_ann}{colname_ann}

\item{gsub_pattern_up}{gsub_pattern_up}

\item{gsub_pattern_down}{gsub_pattern_down}
}
\value{
result
}
\description{
data_integration_preprocess_rmats
}
