% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_integration_preprocess_rmats.R
\name{data_integration_rmats}
\alias{data_integration_rmats}
\title{data_integration_rmats}
\usage{
data_integration_rmats(path_rmats, type = "default", design = NULL)
}
\arguments{
\item{path_rmats}{path_rmats}

\item{type}{type}

\item{design}{design}
}
\value{
result
}
\description{
data_integration_rmats
}
