% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess_expr.R
\name{data_preprocess_expr}
\alias{data_preprocess_expr}
\title{RBP expression matrix pre-preocessing.}
\usage{
data_preprocess_expr(expr, num, cutoff)
}
\arguments{
\item{expr}{RBP expression matrix}

\item{num, cutoff}{Threshold.}
}
\value{
RBP expression matrix after filtering out the low-expression RBPs.
}
\description{
RBP expression matrix pre-preocessing.
}
