% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_integration_preprocess_rmats.R
\name{data_preprocess_rmats}
\alias{data_preprocess_rmats}
\title{data_preprocess_rmats}
\usage{
data_preprocess_rmats(
  data,
  sample_num,
  psi_cutoff,
  selected = NULL,
  reads_cutoff = NULL,
  reads_num = 10,
  colname_ann = FALSE,
  path_rmats = NULL,
  gsub_pattern_up = NULL,
  gsub_pattern_down = NULL
)
}
\arguments{
\item{data}{data}

\item{sample_num}{sample_num}

\item{psi_cutoff}{psi_cutoff}

\item{selected}{if only some samples you want to use}

\item{reads_cutoff}{reads_cutoff}

\item{reads_num}{reads_num}

\item{colname_ann}{colname_ann}

\item{path_rmats}{path_rmats}

\item{gsub_pattern_up}{gsub_pattern_up}

\item{gsub_pattern_down}{gsub_pattern_down}
}
\value{
result
}
\description{
data_preprocess_rmats
}
