% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DS.R
\name{DS}
\alias{DS}
\title{This function provides a method to find the differential splicing events.}
\usage{
DS(x, m, n)
}
\arguments{
\item{x}{a numeric vector about your events splicing level.And x should be ordered by conditions}

\item{m}{number of the first condition.}

\item{n}{number of the second condition.}
}
\value{
The result include p-value, FDR, dPSI and abs_dPSI.
}
\description{
This function provides a method to find the differential splicing events.
}
\examples{
a<-abs(rnorm(10))
DS(a,5,5)


}
