% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DS_matrix.R
\name{DS_matrix}
\alias{DS_matrix}
\title{This function provides methods to find the differential splicing events and construct them into a conforming matrix.}
\usage{
DS_matrix(psi, m, n)
}
\arguments{
\item{psi}{Events splicing matrix（column should be ordered by conditions）}

\item{m}{The number of condition A samples}

\item{n}{The number of condition B samples}
}
\value{
differential splicing matrix
}
\description{
This function provides methods to find the differential splicing events and construct them into a conforming matrix.
}
