% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tab_simple.R
\name{get_direction}
\alias{get_direction}
\title{This function provides a way to get the results in the form of "tab_simple".}
\usage{
get_direction(expr, m, n, result_tab_simple, RBP_cutoff = 0.05)
}
\arguments{
\item{expr}{The path to the file used to store the output.}

\item{m}{Number of the first condition.}

\item{n}{Number of the second condition.}

\item{result_tab_simple}{The output of MRAS.}

\item{RBP_cutoff}{The cutoff of RBP expression between two conditions.}
}
\value{
Return results with direction in "tab_simple" format
}
\description{
This function provides a way to get the results in the form of "tab_simple".
}
