% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DS_mat.R
\name{get_DS_mat}
\alias{get_DS_mat}
\title{This function provides a way to differential splicing matrix.}
\usage{
get_DS_mat(path_use)
}
\arguments{
\item{path_use}{The path to the file used to store the output.}
}
\value{
Get DS_mat file.
}
\description{
This function provides a way to differential splicing matrix.
}
