% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group.R
\name{get_group}
\alias{get_group}
\title{Looking for interacting RBPs.}
\usage{
get_group(rbp_interested)
}
\arguments{
\item{rbp_interested}{The name of the RBP interested.}
}
\value{
interacting RBPs table.
}
\description{
Looking for interacting RBPs.
}
