% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_all.R
\name{get_MRAS_net}
\alias{get_MRAS_net}
\title{get_MRAS_net}
\usage{
get_MRAS_net(
  MRAS_net_group_re,
  rbp_corr_group_work,
  Regulate_threshold = 0.5,
  BS = NULL,
  threads = 2,
  path_use
)
}
\arguments{
\item{MRAS_net_group_re}{first net.}

\item{rbp_corr_group_work}{second net.}

\item{Regulate_threshold}{P cutoff.}

\item{BS}{RBP binding matrix.}

\item{threads}{threads.}

\item{path_use}{The path to the file used to store the output.}
}
\value{
network
}
\description{
get_MRAS_net
}
