% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_all.R
\name{get_MRAS_net_fc}
\alias{get_MRAS_net_fc}
\title{get_MRAS_net_fc}
\usage{
get_MRAS_net_fc(
  rbp_net_mat,
  rbp_corr_work,
  Regulate_threshold = 0.5,
  if_BS = FALSE,
  threads = 2,
  path_use
)
}
\arguments{
\item{rbp_net_mat}{first net.}

\item{rbp_corr_work}{second net.}

\item{Regulate_threshold}{P cutoff.}

\item{if_BS}{if use BS.}

\item{threads}{threads.}

\item{path_use}{The path to the file used to store the output.}
}
\value{
network
}
\description{
get_MRAS_net_fc
}
