% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rbp_event_deal_all_group.R
\name{get_rbp_event_deal_all_group}
\alias{get_rbp_event_deal_all_group}
\title{This function provides a way to obtain the matrix of all RBP-Events regulatory relationships with P > 0.5.}
\usage{
get_rbp_event_deal_all_group(path_use)
}
\arguments{
\item{path_use}{The path to the file used to store the output.}
}
\value{
Get rbp_event_deal_all file and save RData.
        rbp_event_deal_all contains three columns indicating the event ID, the degree to which the RBP regulates it, the RBP
}
\description{
This function provides a way to obtain the matrix of all RBP-Events regulatory relationships with P > 0.5.
}
