% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rbp_event_deal_all_total.R
\name{get_rbp_event_deal_all_total}
\alias{get_rbp_event_deal_all_total}
\title{This function provides a way to obtain the matrix of all RBP-Events regulatory relationships.}
\usage{
get_rbp_event_deal_all_total(path_useful)
}
\arguments{
\item{path_useful}{The path to the file used to store the output.}
}
\value{
Get rbp_event_deal_all file and save RData.
        rbp_event_deal_all_total contains four columns indicating the event ID, the degree of moderation by RBP, the RBP, the degree of difference of splicing events during inference.
}
\description{
This function provides a way to obtain the matrix of all RBP-Events regulatory relationships.
}
