% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_RBP_use.R
\name{get_RBP_use}
\alias{get_RBP_use}
\title{This function provides the means of processing the expression matrix.}
\usage{
get_RBP_use(expr, m, n)
}
\arguments{
\item{expr}{RBP expression matrix.}

\item{m}{Number of the first condition.}

\item{n}{Number of the second condition.}
}
\value{
Preparing for enrichment analysis.
}
\description{
This function provides the means of processing the expression matrix.
}
