% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tab_all.R
\name{get_tab_all}
\alias{get_tab_all}
\title{This function provides a way to get the results in the form of "tab_all".}
\usage{
get_tab_all(path_useful)
}
\arguments{
\item{path_useful}{The path to the file used to store the output.}
}
\value{
Return results in "tab_all" format
}
\description{
This function provides a way to get the results in the form of "tab_all".
}
