% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tab_simple.R
\name{get_tab_simple}
\alias{get_tab_simple}
\title{This function provides a way to get the results in the form of "tab_simple".}
\usage{
get_tab_simple(path_useful)
}
\arguments{
\item{path_useful}{The path to the file used to store the output.}
}
\value{
Return results in "tab_simple" format
}
\description{
This function provides a way to get the results in the form of "tab_simple".
}
