% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Top10.R
\name{get_Top10}
\alias{get_Top10}
\title{This function provides a way to get the results in the form of "Top10".}
\usage{
get_Top10(path_useful)
}
\arguments{
\item{path_useful}{The path to the file used to store the output.}
}
\value{
Return results in "top10" format
}
\description{
This function provides a way to get the results in the form of "Top10".
}
