% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hcc_psi}
\alias{hcc_psi}
\title{some data to test}
\format{
An object of class \code{data.frame} with 4259 rows and 100 columns.
}
\usage{
hcc_psi
}
\description{
some data to test
}
\keyword{a}
\keyword{data}
\keyword{set}
