% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_function.R
\name{id_change}
\alias{id_change}
\title{id_change}
\usage{
id_change(id1, id2, err_len = 2)
}
\arguments{
\item{id1}{Source name needs to be converted.}

\item{id2}{Target name needs to be converted to.}

\item{err_len}{Mismatch length.}
}
\value{
Match the name in id1 to id2.
}
\description{
id_change
}
