% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_function.R
\name{id_find}
\alias{id_find}
\title{id_find}
\usage{
id_find(file_path, software, subtype = NULL)
}
\arguments{
\item{file_path}{the input files path.}

\item{software}{Commonly used software for detecting differential splicing events (rMTAS, JUM, SUPPA).}

\item{subtype}{if `software` = "rMATS",subtype should be "JCEC";if `software` = "JUM",subtype should be "detailed";}
}
\value{
Returns a PSI matrix with canonical splicing event names
}
\description{
id_find
}
