% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_function.R
\name{id_normalization}
\alias{id_normalization}
\title{id_normalization}
\usage{
id_normalization(
  symbol,
  type = "x",
  chr = "x",
  chain = "x",
  loc1 = "x",
  loc2 = "x",
  loc3 = "x",
  loc4 = "x",
  loc5 = "x",
  loc6 = "x",
  loc7 = "x",
  loc8 = "x"
)
}
\arguments{
\item{symbol}{Gene symbol.}

\item{type}{Splicing event type}

\item{chr}{Chromosome number.}

\item{chain}{Positive and negative chains.}

\item{loc1, loc2, loc3, loc4, loc5, loc6, loc7, loc8}{Splicing event coordinates. The unknown position is "x". The user needs to provide key locations or all locations in standard format.}
}
\value{
Standard splicing event name.
}
\description{
id_normalization
}
