% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_all.R
\name{MRAS_net_single_fc_work}
\alias{MRAS_net_single_fc_work}
\title{MRAS_net_single_fc_work}
\usage{
MRAS_net_single_fc_work(
  expr,
  psi,
  fc_mat,
  dpsi_mat,
  num1 = 0.1,
  method = c("pearson", "spearman"),
  sample = batch_matrix,
  dpsi_network_threshold = 0.1,
  BS = NULL,
  threads = 2,
  path_use = path_use
)
}
\arguments{
\item{expr}{RBP expression matrix.}

\item{psi}{Events splicing matrix.}

\item{fc_mat}{fold change matrix.}

\item{dpsi_mat}{dpsi matrix.}

\item{num1}{Denotes the proportion/number of samples to take the expression high or low and to construct the regulatory relationship, respectively.
Input requirements: if the input is less than 1, the input is considered as the sample proportion.
When choosing the input sample proportion, it is required to be no greater than 0.5.}

\item{method}{The method to calculate correlation.}

\item{sample}{batch matrix.}

\item{dpsi_network_threshold}{A threshold value that measures the absolute value of the degree of differential spliicng events in two different condition, above which variable clipping is considered to have occurred, with a default threshold of 0.1}

\item{BS}{if input the binding data.}

\item{threads}{If you want to use multiple threads, you can set threads, where the default parameter is 1.}

\item{path_use}{The path to the file used to store the output.}
}
\value{
network.
}
\description{
MRAS_net_single_fc_work
}
