% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_all.R
\name{MRAS_net_single_sc_work}
\alias{MRAS_net_single_sc_work}
\title{MRAS_net_single_sc_work}
\usage{
MRAS_net_single_sc_work(
  expr,
  psi,
  num1 = 0.5,
  num2 = 0.5,
  method = "spearman",
  dpsi_network_threshold = 0.1,
  BS = NULL,
  threads = 2,
  path_use = path_use
)
}
\arguments{
\item{expr}{RBP expression matrix.}

\item{psi}{Events splicing matrix.}

\item{num1, num2}{Denotes the proportion/number of samples to take the expression high or low and to construct the regulatory relationship, respectively.
Input requirements: if the input is less than 1, the input is considered as the sample proportion, and greater than 1, the input is considered as the number of samples.
When choosing the input sample proportion, it is required to be no greater than 0.5.}

\item{method}{The method to calculate correlation.}

\item{dpsi_network_threshold}{A threshold value that measures the absolute value of the degree of differential spliicng events in two different condition, above which variable clipping is considered to have occurred, with a default threshold of 0.1}

\item{BS}{if input the binding data.}

\item{threads}{If you want to use multiple threads, you can set threads, where the default parameter is 1.}

\item{path_use}{The path to the file used to store the output.}
}
\value{
network.
}
\description{
MRAS_net_single_sc_work
}
