% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRAS_network.R
\name{MRAS_network}
\alias{MRAS_network}
\title{This function constructs the RBP-EVents regulatory relationship.}
\usage{
MRAS_network(
  expr,
  psi,
  num1 = 0.5,
  num2 = 0.5,
  rbp_corr = rbp_corr,
  dpsi_network_threshold = 0.1,
  Regulate_threshold = 0.5,
  threads = 2,
  path_use = path_use
)
}
\arguments{
\item{expr}{RBP expression matrix.}

\item{psi}{Events splicing matrix.}

\item{num1, num2}{Denotes the proportion/number of samples to take the expression high or low and to construct the regulatory relationship, respectively.
Input requirements: if the input is less than 1, the input is considered as the sample proportion, and greater than 1, the input is considered as the number of samples.
When choosing the input sample proportion, it is required to be no greater than 0.5.}

\item{rbp_corr}{The correlation matrix of RBP and events has 4 columns,
the first column is the row name of input x,
the second column is the row name of input y,
the third column is the magnitude of the correlation,
the fourth column is the significance level P-value.}

\item{dpsi_network_threshold}{A threshold value that measures the absolute value of the degree of differential spliicng events in two different condition, above which variable clipping is considered to have occurred, with a default threshold of 0.1}

\item{Regulate_threshold}{A threshold value to measure whether the RBP regulates a splicing event, greater than which the RBP is considered to regulate the splicing event, with a default threshold value of 0.5.}

\item{threads}{If you want to use multiple threads, you can set threads, where the default parameter is 2.}

\item{path_use}{The path to the file used to store the output.}
}
\value{
No return value, the result is stored under the path provided.
}
\description{
This function constructs the RBP-EVents regulatory relationship.
}
