% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psi}
\alias{psi}
\title{somedata to test}
\format{
An object of class \code{data.frame} with 1857 rows and 44 columns.
}
\usage{
psi
}
\description{
somedata to test
}
\keyword{d}
\keyword{dataset}
