% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sc_brca_expr}
\alias{sc_brca_expr}
\title{some data to test}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 268 rows and 515 columns.
}
\usage{
sc_brca_expr
}
\description{
some data to test
}
\keyword{a}
\keyword{data}
\keyword{set}
