% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sc_brca_psi}
\alias{sc_brca_psi}
\title{some data to test}
\format{
An object of class \code{data.frame} with 3197 rows and 515 columns.
}
\usage{
sc_brca_psi
}
\description{
some data to test
}
\keyword{a}
\keyword{data}
\keyword{set}
