% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_matrix.R
\name{score_matrix}
\alias{score_matrix}
\title{This function performs enrichment scoring.}
\usage{
score_matrix(
  rbp_interested,
  Events_DS,
  Event_DS_sig,
  RBP_use,
  DS_pvalue,
  DS_dPSI,
  result_type = c("Top10", "tab_simple", "tab_all"),
  threads = 2,
  rbp_event_deal_all_total,
  rbp_event_deal_all,
  path_use
)
}
\arguments{
\item{rbp_interested}{The name of the RBP interested.}

\item{Events_DS}{The difference matrix of the clipping events, obtained by the MRAS function DS_matrix().}

\item{Event_DS_sig}{Events with significant differences clipped by threshold filtering in Events_DS.}

\item{RBP_use}{RBP expression matrix, the last column is the log2FC value expressed under two conditions}

\item{DS_pvalue}{Significance level of differential splicing analysis.}

\item{DS_dPSI}{Threshold for differential splicing.}

\item{result_type}{The types of return values are "top10", "tab_simple", and "tab_all".
"top10" returns the top 10 MRAS results;
"tab_simple" returns a simplified version of the MRAS scoring matrix (without the events corresponding to the RBP);
"tab_all" returns the full version of the MRAS scoring matrix. matrix (including the events corresponding to the RBP).}

\item{threads}{If you want to use multiple threads, you can set threads, where the default parameter is 2.}

\item{rbp_event_deal_all_total}{Regulatory matrix of RBP-Event.}

\item{rbp_event_deal_all}{Significant regulatory matrix of RBP-Event.}

\item{path_use}{The path to the file used to store the output.}
}
\value{
Results of MRAS analysis.
}
\description{
This function performs enrichment scoring.
}
