% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tj.R
\name{tj}
\alias{tj}
\title{This is a small function for pasting objects together.}
\usage{
tj(name_list, type = ",")
}
\arguments{
\item{name_list}{a vector you want to paste}

\item{type}{which collapse you want to use.If not, then use ","}
}
\value{
a character
}
\description{
This is a small function for pasting objects together.
}
\examples{
a<-c("you","me","he","she","it")
tj(a,"_")
tj(a)
}
