% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group.R
\name{get_group_events}
\alias{get_group_events}
\title{Searching for splicing events co-regulated by interacting RBPs.}
\usage{
get_group_events(rbp_interested, rbp_event_deal_all)
}
\arguments{
\item{rbp_interested}{The name of the RBP interested.}

\item{rbp_event_deal_all}{the network built by MRAS.}
}
\value{
splicing events co-regulated by interacting RBPs.
}
\description{
Searching for splicing events co-regulated by interacting RBPs.
}
