% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRAS.R
\name{MRAS}
\alias{MRAS}
\title{Master Regulator analysis of Alternative Splicing}
\usage{
MRAS(
  input_type,
  expr,
  psi,
  rbp_interested = NULL,
  m = 0,
  n = 0,
  RBP_cutoff = 0.05,
  DS_pvalue = 0.05,
  DS_dPSI = 0.1,
  rbp_event_deal_all_total = NULL,
  rbp_event_deal_all = NULL,
  method = "spearman",
  BS = NULL,
  group = T,
  dpsi_network_threshold = 0.1,
  Regulate_threshold = 0.5,
  rbp_net_mat_group = NULL,
  num1 = 0.5,
  num2 = 0.5,
  sc = F,
  threads = 1,
  path_use,
  smooth = FALSE,
  result_type = c("Top10", "tab_simple", "tab_all")
)
}
\arguments{
\item{input_type}{input type, different options for user's different needs.
"1": input differential splicing events and use pre-constructed regulatory network;
"2": input data and inferred relationships by MRAS.
"3": input data and construct a network using the user's own data}

\item{expr}{RBP expression matrix.}

\item{psi}{Events splicing matrix.}

\item{rbp_interested}{The name of the RBP interested.}

\item{m}{Number of the first condition.}

\item{n}{Number of the second condition.}

\item{RBP_cutoff}{The cutoff of RBP expression between two conditions.}

\item{DS_pvalue}{Significance level of differential splicing analysis.}

\item{DS_dPSI}{Threshold for differential splicing.}

\item{rbp_event_deal_all_total, rbp_event_deal_all}{the network built by MRAS.}

\item{method}{Method of finding correlation.}

\item{BS}{RBP binding matrix.}

\item{group}{TRUE or FALSE.If consider the synergy between RBPs.}

\item{dpsi_network_threshold}{A threshold value that measures the absolute value of the degree of differential spliicng events in two different condition, above which variable clipping is considered to have occurred, with a default threshold of 0.1}

\item{Regulate_threshold}{A threshold value to measure whether the RBP regulates a splicing event, greater than which the RBP is considered to regulate the splicing event, with a default threshold value of 0.5.}

\item{rbp_net_mat_group}{first net which can get from "MRAS".}

\item{num1, num2}{Denotes the proportion/number of samples to take the expression high or low and to construct the regulatory relationship, respectively.
Input requirements: if the input is less than 1, the input is considered as the sample proportion, and greater than 1, the input is considered as the number of samples.
When choosing the input sample proportion, it is required to be no greater than 0.5.}

\item{sc}{TRUE or FALSE. If single cell data.}

\item{threads}{If you want to use multiple threads, you can set threads, where the default parameter is 1.}

\item{path_use}{The path to the file used to store the output.}

\item{smooth}{network building type.User can choose "TRUE" or "FALSE" in the third input type.}

\item{result_type}{The types of return values are "top10", "tab_simple", and "tab_all".
"top10" returns the top 10 MRAS results;
"tab_simple" returns a simplified version of the MRAS scoring matrix (without the events corresponding to the RBP);
"tab_all" returns the full version of the MRAS scoring matrix. matrix (including the events corresponding to the RBP).}
}
\value{
MRAS result. If you have any questions,you can get help here (https://github.com/zhou-lei5/MRAS).
}
\description{
Master Regulator analysis of Alternative Splicing
}
